Pivot from pytorch to tensorflow
This commit switches from using pytorch, which was never actually setup
to using tensorflow. This is mostly because I was able to figure out
what to do with the tensorflow docs and was still lost on how to process
the data in pytorch. The tensorflow project is also a bit more mature
and also actually packaged on pypi making it easier to integrate into
requirements.

This starts by using the prebuilt DNNClassifier to process the data,
we'll still need to tune this to see if we can get good results from it.
We might have to change the model we're using or create our own. But
this was a starting point. Still missing is the export model
functionality which we'll need to add before we can start predicting.