Fix type issue on python 3.5
In python 3.6 support for binary input to json.loads was added. It
assumes binary data is utf8, utf16, or utf32 encoded and just decodes
the input. However, earlier versions of python don't have this feature
in the stdlib json module. This commit adjusts the call for that and
manually decodes the mqtt payload before running it through the json
parser.