Implement a DB trainer
When pulling data from the DB we get runs by name, and then get
a result object for each matching run. The basic trainer is ready
but there are a few things missing still:

- caching and local lookup of dstat files
- the gather_results module will have to be refactored to do some
  smarter session mgtm
- actually do the training

We may also consider caching the other metadata that we get from
the DB so that we have a fully local dataset.