Fix length normalization and add vectorization
This commit fixes the previously added length normalization. In my
testing using append() on a dataframe or a series didn't do what I
expected initially. At the same time this also adds vectorization to the
dataframe. So instead of a big 5500x31 dataframe it returns 2 vectors of
170500 length. The first is all the values and the second is the labels
of each type of data (just in case we need it)