Use normalization to prepare the data
In the local trainer, build a vector of examples and a vector of
classes. Examples are test runs i.e. a dstat file, classes are the
test run result, i.e. pass of fail (0 or 1).

Start using TF / ML nomenclature in variable names a bit more to
get less confused about what's what.