Continue work on SVM trainer
This commit picks up where the previous left off getting the SVM trainer
WIP closer to working. This fixes several issues encountered when trying
to use the SVM trainer. It's not really 100% there yet, but it's closer.
It's also worth noting since I haven't figured out how to use the
dataset storing functionality I've been doing all my testing with the
MQTT trainer. The call paths in ciml.gather_results seem to be a bit
confusing and this might trample on a call path for one of the other 3
trainer CLIs.

This is failing right now on passing the input data to the SVM model.
Things are in the wrong format still, it's failing with the the
following error:

ValueError: inputs must be a list of at least one Tensor with the same
dtype and shape