Add more tf logging and add option to force GPU usage
In the investigation on why my GPU wasn't being leveraged for model
training additional logging was needed to see devices detected and how
they were being used. At the same time this commit adds an option to try
and force GPU usage for the model training. As of right now when using
force GPU the program will be loaded onto the GPU. (which it doesn't
without) However, it appears to just sit there idle eating a small
amount of GPU memory. More investigation is required, but this a start.