Add feature normalization
Normalize each feature using the normalizer function passed to
a real_valued_column feature.

Bring back the computation of the desired example duration after
re-sample, which is based on the value of the desired example
duration in seconds reshaped with the resampling.

Also close figures, to stop warnings about memory utilization.