Fix model evaluation call
This commit fixes the model evaluation call, we were never setting a
step count and it was sitting there iterating through the model (or
doing something) without ever exiting. We really only need a single step
to load the model and pull the loss and accuracy statistics from it. By
only evaluating a single step the training function will exit and we can
continue developing our model usage.