Start implementation of mqtt predicter command
This commit adds the initial groundwork for a mqtt predicter. This will
listen to the mqtt stream for database add events, and when one of the
proper build name is received it will use the dstat data to run
predict() on the model with that input data.

This hasn't really been tested yet, but at least it doesn't crash when
the command starts and begins waiting for a proper event. I expect the
predict() call path will need work, because I was just shooting in the
dark. This is just laying the groundwork, we can fix it as we test
things. Having a local and/or db predicition path would likely speed up
testing and enable us to verify the prediction tensorflow path.