Add logging to ciml-mqtt-predict
This commit adds logging to the ciml-mqtt-predict process. Prior to this
commit there was no indication of what the process was doing. This adds
2 levels the default will print when an event is processed, and whether
it downloaded data or not. The debug level which is triggered with the
'--debug' flag will print more details like when the main loop starts
and when incoming mqtt events are processed.