Add ciml-db-predict command
Similar to the ciml-mqtt-predict command, this enables running a model
prediction based on dstat files pulled from the db. But instead of
waiting for an mqtt event triggering the prediction this takes in a zuul
build uuid as an argument. This is used to pull in the dstat file and
check that against the model.

Right now the prediction code isn't working yet, because the input
function and data format aren't quite right. But the call path setting
things up appears to work, which speeds up the feedback loop for
testing.