Move dstat files to a central cache
There's no need for every dataset to re-download dstat files.
Besides dstat file often expire on the logs server, so it's important
to preserve a local copy.

Move the current folder structure:

- dataset1
   \- raw
       |- run_uuid1.csv.gz
       \- run_uuid1.json.gz
- dataset2
   \- raw
       |- run_uuid1.csv.gz
       |- run_uuid1.json.gz
       |- run_uuid2.csv.gz
       \- run_uuid2.json.gz

To a new folder structure:

- .raw
    |- run_uuid1.csv.gz
    \- run_uuid2.csv.gz
- .metadata
    |- run_uuid1.json.gz
    \- run_uuid2.json.gz
- dataset1
   \- runs.json.gz (list of UUIDs)
- dataset2
   \- runs.json.gz (list of UUIDs)

This is not backward compatible, so with the new code existing dataset
will fail to load properly unless files are moved around.
From the data folder:

$ mkdir .raw
$ find . -name '*.csv.gz' -exec cp '{}' .raw \;

Metadata files can be re-built from the DB, this will happen
automatically if the files cannot be found.

This will also leave .*.gz files lying around in the metadata folders
but this is done on purpose so that migration can be checked before
files are deleted. Deletion can be done with:

$ find . -name raw -type d -exec rm -rf '{}' \;