Start work on using a DNN classifier
This commit lays the initial groundwork for using a deep neural network
to try and classify failures from the dstat data. Instead of unrolling
the data into a linear series this keeps the dataframe as a matrix with
a feature for each column. Also instead of treating the status as a
vector with all 0s or 1s based on whether the run passed or failed the
DNN classifier attempts to match the first test failure timing. So
on a failed run it will be all 0s until the first test returns a failure
and then after that it will be 1s until the end.