Add option to perform DNN training
This commit adds a method to perform training with the DNN model. By
leveraging the --estimator argument on ciml-build-dataset and setting it
to anything besides the default value when a run is acquired it will
also use that run to train the DNN model. This isn't an ideal solution
because it differs from the normal workflow used for the local training.
It was used because the cached data wasn't including the tests
information which is required to build the classification vector. This
was implemented as a workaround until we can fix the caching.