Handle missing dstat data
There is an edge case where the artifact link is missing on processing
cached results. I haven't figured out where exactly this is coming from.
But to make this clear this commit handles this case to get the run
uuids so that we can remove the missing runs from the dataset input.