Fix pep8 issues
This commit goes through and cleans up all the issues reported by flake8
on the repo. We really should automate this, but since we're still not
using code review for changes there isn't a good mechanism to force this
(short of a git commit hook or something). So periodically cleaning
things up by hand is the best we can do.