Tweak hidden layers in DNN model
This commit tweaks the hidden layer setup in the DNN model a bit. This
is more just an experiment to try different model setups to get a
better feel for what works. From the initial testing it looks like this
is working a bit better than the previous setup, on average the model
loss is less than before.