Add CLI path to call DNN model evaluate
This commit adds a --evaluate option on the db_trainer function that
is used to run model evaluation on the DNN path. It leverages the same
functionality as the training path, but instead calls evaluate instead
of train. In the long run this probably isn't ideal and we'll want a
better path for using the DNN, but for now this works well enough.