Mix results between failures and passes
When building a dataset (or training a DNN) we were getting a large
number of successes relative to the failures. This is because in general
things normally pass (which is part of why we're doing this whole
exercise in using ML). But this has the effect the model may be biased
and not actually have a representative set of failures. To address this
case this commit reorders the data it gets from the database. It'll now
evenly mix the runs between failures and successful runs up front and
then have the remaining successful runs at the end.