Update ubuntu version in ciml-base image
Ciml depends on python >=3.6 and we need to make sure we're running with
that. However our ciml-base image was being deployed with ubuntu:latest
which as of right now is based on 16.04 xenial which only packages
python 3.5. To fix this (at least until 18.04 bionic is released) this
changes the ubuntu version to be ubuntu:rolling which pulls in 17.10
artful. This will get us the correct version of python.