Add REST API service for model training
This commit starts the process of adding a REST API service for running
model training. This will give a convenient method of training (or
re-training) a model that is being shared and used by another
process/daemon. The idea here is that this server will have shared
storage with the prediction daemon and they'll run in conjuction so we
can manually trigger training and have that automatically reflected in
the prediction.

Right now this server has the basic framework for the API and a single
method of training the model for a single test_name. More parameters and
methods of training will need to be added in the future. Additionally,
while the API server will start up fine and respond to requests, there
are still some issues with the caching and data storage that need to be
resolved before the training is actually run.