Use model_dir api config option for data cache too
This commit adds support to the data caching in the training rest api to
choose where the data is stored. Previously it was hard coded to a path
relative to the code repo. However when running this in a real
deployment you'll need to put that on shared storage. Right now it uses
the same model_dir directory from the config to specify this, in the
future we might need to separate that to a separate option.