Fix SVM training api
This commit fixes a couple of issues with the training api so now that
it will properly train on a POST call.

The first issue fixed was with the runs caching. We were properly
caching the run uuid list and on missing data trimming it. But on the
second run we were unconditionally overwriting that trimmed data set.
This caused all calls to always fail because we were never using the
trimmed data.

The second issue fixed here was the sync nature of the API. Training is
a long running task. Previously the API was naive and just ran the
training and wouldn't return until it finished. However this was almost
guaranteed to cause a response timeout before the training finished. To
address this the API call is now async. A thread will be spawned to
run the actual training and once that's launched it will return a 202.