Add initial attempt at api server deployment
This commit adds the basic framework for running an api server in
another container on the model pod. This includes the docker image for
that container. Right now it doesn't quite work the docker image isn't
starting the processes properly. But, it's close enough that we can
iterate on that in future commits.