Fix aspects of ciml-api deploy
This commit fixes several factors related to the ciml-api deployment.
It's still not 100% there, but at least now both nginx and uwsgi are
running in the container. The piece that's still missing is the nginx
configuration to get it to use the uwsgi proxy for all incoming requests
off of / instead of the default sample configuration, which just has a
static html welcome page.