Finalize ciml-api image scaffolding
This commit finally fixes the last deployment issues with the basic
framework for running the container for the ciml-api. With this the
container will properly run nginx and uwsgi with nginx configured to
proxy the requests to uwsgi running the ciml-api server.

Note while the web server and uwsgi server are running properly the
ciml-api isn't quite working. While it will accept the incoming requests
there are issues with the db session handling which cause it to fail.
Additionally there is a lot proxying going on and either the k8s load
balancer or nginx is holding the connection open too long which causes
a 502 response. Even though the sole training API method should return
quickly that 202 response isn't being passed up through the proxies
correctly.