Improve dataset building
We only use the db_trainer for dataset building, so renaming it,
and changing it to use a new gather_data module function that loops
through runs re-using a single DB session.

Store on disk under .unavailable the ID of runs for which no dstat
data is available to download (nor cached), so that we don't check
them everytime using an expensive HTTP call.

Handle CSV parsing issues, delete a cache file when they come from
a local cache, since this usally means that the local cache is
corrupt).