Add support for experiments and DNN
Workflow now includes:
- create an experiment for a dataset
- train based on dataset/experiment

Introduced a tf_trainer class with helpers to setup the training.
Not using the class model anymore with a class encapsulating on
estimator.

Fixed the input_fn to slice the data by feature.

There is still a lot of cruft in the repo, code that is not
actually used anymore, nomenclature that is not aligned,
this can be fixed in a series of commit later on.