Fix data normalization
The dev/test/training split algorithm was wrong, fixing it.
Support now also have sets of size 0.

Cleanup and NaN coming out of resampling.

Fix the normalization formula to handle the case of a single
example or a flat feature (max-min==0).

Enhance the ability to select examples when building a dataset.
Instead of limit support slices.