Complete s3 support for raw data
With s3 data possibly stored on s3, the build-dataset command must
also be able to deal with s3.

Since boto3 only allows access to exception classes via the client
object (not class), we must always have a fall-back s3 client
even when we're not using s3, just to be able to run the code.

For node_provider, reduce the number of classes to 6, by putting
all rax and all ovh providers together (respectively).