Extend s3 support to datasets def and data
Allow for dataset definition and normalized data to be store in s3
as well. This normalized data can act as input to a cloud training
system such as kubeflow, ffdl or ciml own.

In gather_results, if data_path is in s3:// format and the s3 client
is not passed, raise a ValueError.