Bump minimum sqlalchemy version
The Github UI is being annoying and saying we have a security
vulnerability in the project because our minimum sqlalchmey requirement
was below 1.3.0. That's because there were 2 vulnerabilities
CVE-2019-7164 and CVE-2019-7548 (which were both sql injection
vulnerabilities). While it's not likely that ciml will have any exposure
because of these, and that we don't ever install the minimum version
listed in the requirements file when using it this commit bumps the
minimum so the annoying alert on the UI goes away.