Add support for aggregation functions
Add an extra parameter to the create_dataset API, that allows
to specify a list of aggregation functions by name.
The name may refer to a numpy function or a function defined in
the trainer.py module.

When aggregation functions are specified, no truncation nor
padding takes place. Each aggregation function is executed
instead with a sample as input.
Sampling and feature selection behave the same way.