Use a third dataset for predictions
CIML supports splitting the dataset in three parts. We never really
used the "dev" since there was no need for a smaller dataset for
faster iterations.

We want to get predictions from the estimator so that we may use
external tools to build the confusion matrix and draw further
metrics. We can use the "dev" set for that.