Use tensorflow v1 compat API
Tensorflow provides a v1 compat API that can be used to run
v1 APIs with a TF2 installed. Switch all of our import to
those makes it possible to lift the tensorflow pin while
having CIML code still working.

There is no TF<2 for python 3.8, so this allows running
CIML on latest versions of python.